+++ TMG/q_shared.c
@@ -1337,7 +1337,7 @@ void Info_SetValueForKey (char *s, char *key, char *value)
 {
 	char	newi[MAX_INFO_STRING], *v;
 	int		c;
-	size_t	maxsize = MAX_INFO_STRING;
+	int		maxsize = MAX_INFO_STRING;
 
 	if (strstr (key, "\\") || strstr (value, "\\") )
 	{
